DO NOT SUBMIT [ELF] Emit compact unwind information in .eh_frame
This adds support for compact unwind descriptors in ELF .eh_frame
sections, extending the existing Mach-O assembler implementation.
Key changes:

**X86 backend**
- Generalize the Mach-O compact unwind code to work for ELF targets.
- Add option "epilog-cfi" to allow disabling epilog CFI (disabled for
  Darwin; see https://reviews.llvm.org/D42848). The current compact unwind
  code does not handle `popq %rbp; .cfi_def_cfa %rsp, 8; ret`

**Integrated assembler**
- Emit .eh_frame CIEs with augmentation character 'C' for compact unwind
- Replace FDE instructions with 64-bit unwind descriptors (extended from
  Mach-O's 32-bit format to support future asynchronous unwinding)
- The augmentation character 'C' mechanism should be considered
  experimental. As a prior art, MIPS compact exception tables uses a
  different section `.eh_frame_entry` instead.

**Linker (lld)**
- Split FDEs into two groups: descriptor-based (augmentation 'C') and
  instruction-based
- Generate .eh_frame_hdr version 2 with 12-byte table entries when compact
  FDEs are present: (pc_ptr, unwind_descriptor_or_fde_ptr)
- TODO: Place LSDA in a separate storage

**Other tools**
- llvm-readelf --unwind: Dump unwind descriptors in .eh_frame_hdr and
  .eh_frame
- llvm-dwarfdump --eh-frame: Recognize augmentation character 'C' in .eh_frame CIEs

Note: libunwind ELF support are planned for future work.

+++ lld/ELF/InputSection.h
@@ -380,7 +380,7 @@ struct EhSectionPiece {
   }
 
   size_t inputOff;
-  ssize_t outputOff = -1;
+  int64_t outputOff = -1;
   InputSectionBase *sec;
   uint32_t size;
   unsigned firstRelocation;
